% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_prep_climate_data.R
\name{calc_clim_drought_period}
\alias{calc_clim_drought_period}
\title{Calculate Climate Drought Period}
\usage{
calc_clim_drought_period(
  clim_data,
  growth_period = 12,
  growth_end = c(NH = 8, SH = 2),
  spei_scale = 1,
  rescale_spei = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{clim_data}{A data.table containing climate data with columns: Id (e.g. can001), Year, Month, TAve, Prec, Lat (latitude).}

\item{growth_period}{Numeric, the number of months prior to \code{growth_end} that are thought to have influence growth. Will be used to average climate conditions for that period. E.g. if northern hemisphere summer (Jun, Jul and Aug) is the interest of investigation then this value would be 3 with \code{growth_end = 8} (for August)}

\item{growth_end}{Numeric, the month when growth is thought to have stopped (default = \code{8} for Northern Hemisphere and \code{2} for Southern Hemisphere).}

\item{spei_scale}{Numeric, the scale at which SPEI is calculated (default = 1 month). Refer to the time scale over which water deficit accumulates. A value of 3 indicates that the SPEI for March is based on the accumulated water deficit over January, February and March.}

\item{rescale_spei}{Logical, whether to rescale SPEI averaged for the \code{growth_period}.}

\item{verbose}{Logical. If TRUE (default), prints a summary of missing}
}
\value{
A data.table with adjusted climate data, including calculated drought years, PET, water balance, and SPEI.
}
\description{
This function adjusts climate data to define drought years, shifting months appropriately for the Northern and Southern Hemispheres.
It also computes PET, water balance, and SPEI.
}
