% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{chron_itrdb_meta}
\alias{chron_itrdb_meta}
\title{Chronology metadata}
\format{
A data frame with 29 columns (including Id, site information, coordinates, species, etc.).
}
\source{
Extracted from the official ITRDB metadata using \code{load_thesis_data().
}
\usage{
chron_itrdb_meta
}
\description{
Metadata associated with the ITRDB chronologies.
}
\keyword{datasets}
