% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{clusters}
\alias{clusters}
\title{Cluster grouping data}
\format{
A data frame with columns:
\describe{
\item{Id}{Unique site identifier.}
\item{group_col}{Continent delineation + cluster code.}
}
}
\source{
Derived from cluster analyses using Partition Around Medoids (PAM) done
seperately by each continent delineation and using elbow method for
identifying ideal number clusters per continent delineation.
Derived from \code{load_thesis_data()
}
\usage{
clusters
}
\description{
Group identifiers for each chronology site, used to set \code{chron_group_col}.
}
\keyword{datasets}
