% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05a_helpers_model_resilience_indices.R
\name{helper_nls_compute_intersections}
\alias{helper_nls_compute_intersections}
\title{Compute Intersection with Full Resilience}
\usage{
helper_nls_compute_intersections(ci_data, model_successful)
}
\arguments{
\item{ci_data}{A data.table with CI bands and full resilience values.}

\item{model_successfull}{Logical indicates whether and nls converged.}
}
\value{
A list with intersection types and threshold Resistance values.
}
\description{
Computes where model CI intersects full resilience line.
}
