% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05a_helpers_model_resilience_indices.R
\name{helper_nls_fit_recovery_model}
\alias{helper_nls_fit_recovery_model}
\title{Fit Recovery Model}
\usage{
helper_nls_fit_recovery_model(data)
}
\arguments{
\item{data}{A data.table or data.frame with columns Resistance and Recovery.}
}
\value{
A fitted \code{nls} model object or a character message if error.
}
\description{
Fits a negative exponential model: Recovery ~ z * Resistance^b.
}
