% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05a_helpers_model_resilience_indices.R
\name{helper_nls_predict_ci_band}
\alias{helper_nls_predict_ci_band}
\title{Predict Confidence Bands (Revised)}
\usage{
helper_nls_predict_ci_band(
  model_successful,
  model,
  nls_boot,
  data,
  id,
  idx,
  total,
  start_time,
  verbose = TRUE
)
}
\arguments{
\item{model}{A fitted \code{nls} model object.}

\item{nls_boot}{A \code{nlsBoot} object (bootstrapped model).}

\item{data}{A data.table with a numeric column \code{Resistance}.}

\item{id}{Character File id to be processed.}

\item{idx}{Character Index of File id to be processed.}

\item{total}{POSIXct of the start time of the function.}

\item{start_time}{POSIXct of the start time of the file id to be processed.}

\item{verbose}{Logical outputs messages in the console and in a log file to current directory.}

\item{model_successfull}{Logical indicates whether and nls converged.}
}
\value{
A data.table with Resistance, median_ci, lwr_ci, upr_ci, fit_sp_ci, and full_res.
}
\description{
Predicts median, lower, and upper CI bands using nlsBoot,
automatically deriving the resistance range from the supplied data.
}
