% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05a_helpers_model_resillience_indices.R
\name{helper_nls_range_limits}
\alias{helper_nls_range_limits}
\title{Determine Recovery Range Limits}
\usage{
helper_nls_range_limits(
  data,
  upr_cross_type,
  upr_intsct_thr,
  lwr_cross_type,
  lwr_intsct_thr
)
}
\arguments{
\item{data}{A data.table containing Resistance values.}

\item{upr_cross_type}{String, type of upper CI intersection.}

\item{upr_intsct_thr}{Numeric, upper intersection threshold.}

\item{lwr_cross_type}{String, type of lower CI intersection.}

\item{lwr_intsct_thr}{Numeric, lower intersection threshold.}
}
\value{
A list with named limits: unde_rec_lower_limit, unde_rec_upper_limit,
full_rec_lower_limit, full_rec_upper_limit,
over_rec_lower_limit, over_rec_upper_limit.
}
\description{
Computes lower and upper resistance limits for under-, full-, and over-recovery
based on intersection thresholds with the full resilience line.
}
