% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_prep_chronology_data.R
\name{impute_chronology_data}
\alias{impute_chronology_data}
\title{Impute Missing Tree-Ring Chronology Data}
\usage{
impute_chronology_data(chron_data, verbose = TRUE)
}
\arguments{
\item{chron_data}{A \code{data.frame} (or \code{data.table}) containing tree-ring chronology data with the following required columns:
\itemize{
\item \code{"Id"}: Unique identifier for each site.
\item \code{"Year"}: Year of measurement.
\item At least one of \code{"RWI"} (Ring Width Index) or \code{"RES"} (Residual chronology from an autoregressive modelling).
}}

\item{verbose}{Logical. Whether to display progress messages (default = TRUE).}
}
\value{
A \code{data.table} with missing values imputed using Random Forest.
}
\description{
This function performs imputation on missing values in tree-ring width index (RWI) or residual (RES) chronology data.
It utilizes Random Forest imputation (\code{rfImpute}) to fill gaps based on available data.
}
\details{
The function reshapes the data from long to wide format, where years become columns.
It then applies \code{rfImpute} from the \code{randomForest} package to estimate missing values based on other available data.
The function will use all data available for imputation. User must use with care.
}
\examples{
library(data.table)
chron_data <- data.table(
  Id = rep(1:3, each = 5),
  Year = rep(2000:2004, times = 3),
  RWI = c(NA, 1.2, 1.1, 1.3, 1.2, 2.1, NA, 2.3, 2.4, 2.2, 1.5, 1.6, NA, 1.8, 1.9),
  RES = c(0.2, NA, 0.1, 0.3, 0.2, 0.5, 0.6, 0.4, 0.3, 0.2, NA, 0.8, 0.7, 0.6, 0.5)
)

imputed_data <- impute_chronology_data(chron_data)
print(imputed_data)

}
