% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{std_drought_clus}
\alias{std_drought_clus}
\title{Cluster grouping data}
\format{
A data frame with columns:
\describe{
\item{Id}{Unique site identifier.}
\item{CLUSTER2}{Continuous integer numbering for all clusters in the dataset.}
\item{CLUSTER3}{Continuous integer numbering for all cluster included in the analysis. i.e. CLUSTER3_STATUS == "Included"}
\item{CLUSTER3_STATUS}{Character of included or not in the analysis.}
\item{COLOR}{Cluster color}
\item{name}{Region name for easier interpretation.}
\item{group_col}{Continent delineation + cluster code.}
}
}
\source{
Derived from cluster analyses using Partition Around Medoids (PAM) done
separately by each continent delineation and using elbow method for
identifying ideal number clusters per continent delineation.
Derived from \code{load_thesis_data()
}
\usage{
std_drought_clus
}
\description{
Group identifiers for each chronology site, used to set \code{chron_group_col}.
}
\keyword{datasets}
