% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05a_helpers_model_resillience_indices.R
\name{helper_nls_bootstrap_nls_model}
\alias{helper_nls_bootstrap_nls_model}
\title{Bootstrap NLS Model}
\usage{
helper_nls_bootstrap_nls_model(model, data, id)
}
\arguments{
\item{model}{A fitted \code{nls} object.}

\item{data}{The original dataset used to fit the model.}
}
\value{
A bootstrapped \code{nlsBoot} object or NA if error.
}
\description{
Bootstraps a fitted \code{nls} model using \code{nlstools::nlsBoot}.
}
