% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05a_helpers_model_resillience_indices.R
\name{helper_nls_compute_intersections}
\alias{helper_nls_compute_intersections}
\title{Compute Intersection with Full Resilience}
\usage{
helper_nls_compute_intersections(ci_data, modeled_successful)
}
\arguments{
\item{ci_data}{A data.table with CI bands and full resilience values.}
}
\value{
A list with intersection types and threshold Resistance values.
}
\description{
Computes where model CI intersects full resilience line.
}
